Update patch set 1
Patch Set 1: Code-Review-1

(6 comments)

All replacements of catch(Throwable t) to catch(Exception t) should be changed to catch(Exception e). 

But as far as catching generic Exception, I think these should all be narrowed but this seems to be endemic in the code, so I'll leave to others.

Patch-set: 1
Label: Code-Review=-1