Update patch set 4
Patch Set 4: Code-Review+1

Tabs and spaces seem endemic to underlying code and not to the changes proposed here. Agree with Skip on that count. Otherwise seem fine.

It would be good to have header of next patch be "Update ASDC References to SDC 2/2" with body explaining in similar way as this commit message.

Patch-set: 4