Update patch set 2
Patch Set 2: Code-Review-2

(2 comments)

Please rework this change, it does not address the Sonar violation and creates a new one.  A SvcLogicException is never thrown from the try so it should not be caught.    All of the processing in the catch block for the Exception has been removed and needs to be put back.

Addressing the "log or rethrow" violation should be addresses by logging the exception in the original catch, e.g.,

logger.error("Failed in getVnfInfo ", e);

Patch-set: 2
Reviewer: Gerrit User 1125 <1125@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2