Update patch set 2
Patch Set 2:

Hi Michal.   Changes look good except for one question about the method/property re-ordering.   Is the re-ordering necessary per the java style guidelines or for the readability of the code?   If not, I think the original order should be maintained and reduce the churn in the change logs.

Patch-set: 2
Label: Code-Review=0