Update patch set 2
Patch Set 2:

Hi Jim, This is my answer to your concern about ConfigComponentAdaptor changes. The logic has been added to make the adapter more complient to NETCONF standard while not breaking the existing functionality which has been tested with various VNFs. As per NETCONF standard  every running config should have <data>..</data> .So any lines with in this two tags will be considered as the running config coming from the VNF . Where as earlier logic was looking for <configuration>...</configuration>. All lines including <configuration>...</configuration> tags are considered as running configuration. In case of some VNFs,it has been found that <configuration>...</configuration> tags are missing in the running config.So logic has been upgraded so that the adapter will consider both <configuration>...</configuration> and <data>..</data> while parsing the running config . It is backward compatible as well as in line with NETCONF standard .

Patch-set: 2