Update patch set 1
Patch Set 1: Code-Review-1

We have asked the original authors of the code, and they are not able to confirm that this change will not cause any problems with existing functionality.

The change will need to be tested with a vnf of this protocol to make sure nothing breaks.

Patch-set: 1
Reviewer: Gerrit User 49 <49@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1