Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I don't think this change is needed - doesn't provide any value and requires additional code to be updated should someone decide to change one of the constants in the future.  If the whole point of this test is to increase code coverage in tests, that should happen automatically when the code that USES these constants is tested.

Patch-set: 1
Label: Code-Review=-1