Update patch set 1
Patch Set 1:

> i would like to know if constants.java is already getting covered
 > through another class in code coverage?

it should be, but whether it is or not is kind of irrelevant.  What's the point of testing that a constant value equals the constant value (i.e., itself)?  That just means that the test code has to be updated any time someone updates the constants - it just doesn't seem to add any value.

Patch-set: 1