Update patch set 1
Patch Set 1:

Hello Takamune,
Basically the goal of this change was to have the LicenseManagerBean class fully tested also replace string-based xml parsing with a proper xml parsing lib (stax in this case). The unit tests also use a proper xml as the input to test this. The other refactor mostly involved extracting the xml parsing logic and stream handling into separate classes. This change was necessary because unfortunately it was otherwise impossible to cover all outcomes (like throwing parser exceptions). 
I am sorry if some of these changes haven't been reflected in the jira. Would you like us to update the jira with all necessary changes or create a new one for this ?

Patch-set: 1