Update patch set 1
Patch Set 1:

Thank you very much for the review, Takamune and Jim. I updated JIRA with a more detailed description of the changes. It's still a draft, but it's good that you posted your comments already. I'll wait for the last review from Michal and correct all the bugs together.

Patch-set: 1