Update patch set 1
Patch Set 1:

> i hhave added the test case to cover the exception scenario itself

I mean, is the method declared to throw a general Exception?  If so, then the test is fine as is.  However, if the method is declared to throw a specific type of exception (e.g., IOException), then change the test to "expected = Xxx.class", where "Xxx" is the specific type that the method throws.

Patch-set: 1