refactored code in ExecuteNodeActionImpl.java
created a method a called getVserverRelations(SvcLogicContext vnfCtx,
SvcLogicContext ctx) having the logic for looping through relationship-list
data, to get vserver relations. This piece of code was earlier not
covered by test cases. By putting this logic in a separate method, we
will be able to write a specific test case for this logic and will be
able to increase the code coverage by a big amount.

Issue-ID: APPC-1086
Change-Id: Iffa2a4f04a1fd2b93edc4bcdfef8625f196a305e
Signed-off-by: Sandeep J <sandeejh@in.ibm.com>