Update patch set 4
Patch Set 4:

Please add the test result for verification as well. The screen shot of the api doc where the LCM yang is compiled as well as hosted in ODL successfully and exposed as service endpoint. Each service endpoint should be expandable and able to process any lcm request. A negative test or regression test of existing functionality will be helpful to determine these changes are good to merge and has no negative impact to the existing LCM functionality.

Patch-set: 4