Update patch set 1
Patch Set 1: Code-Review-1

Yes, I agree. In the Policy Framework, we encourage people to fix sonar issues. However, in order to cut down on the number of reviews, to make the process more efficient, and to make the maximum use of people's time, we ask people to package all the reviews for a package into a single review.

While we don't want reviews to get too big, certainly it would make sense to have reviews with 10 or 15 files changed if all that's changing is removing unused imports.

Patch-set: 1