Update patch set 1
Patch Set 1:

The tests seem to be running in 1.1 seconds for me on Eclipse/Windows although I can see from the Jenkins logs that it is over 3s there. I tried mocking out the RequestContext.isAlive() call but it didn't seem to make any difference to the timing.
The RequestContextTest class in the same package was running over 40 seconds in Jenkins build so there are obviously some time savings to be made in this part of the test code.

Patch-set: 1
Reviewer: Gerrit User 3979 <3979@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0