Update patch set 2
Patch Set 2:

It is not an abstract class but it has a protected constructor.
There is an example of how to instantiate a X509Certificate object in javadocs description of the class and example .PEM files to instantiate a certificate are available online.

Patch-set: 2
Reviewer: Gerrit User 3979 <3979@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0