Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I'm not sure if this change will reduce the number of Sonar issues. I think if we declare the classes Serializable that Sonar will be asking for a serialization uid field. I think it is better to ignore this Sonar issue as the fields are JSON properties and are sufficiently serializable as such.

Patch-set: 1
Reviewer: Gerrit User 3979 <3979@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1