Escape quotes in json strings
The JSONObject parser strips all text of backslash escape characters, even if it is
treaing the object as a string and not a series of json keys. If a string contains a
json object within it, the removal of these escape characters causes problems. Adding
the JSONObject.quote method on all strings to automatically add escape characters to
any strings which contain characters requiring escapes.

Change-Id: Ifac68844dd132971ba35803f1b9c03a5372bd500
Signed-off-by: Patrick Brady <patrick.brady@att.com>
Issue-ID: APPC-1576