Fix eelf resource not found
The eelf package needs to be exported from appc-common
where the resource bundle is added. Doing this was
causing karaf to crash on install of artifact-handler.
The cause was all of the classes being exported by the
maven shade plugin of appc-yang-generator. Removing the
shade plugin seemed to have no bad consequences.

Change-Id: Ic0758b342b8109e28ec8e11a48b2693cea7b56db
Signed-off-by: Patrick Brady <patrick.brady@att.com>
Issue-ID: APPC-1713