Feature for micro service communication
The appc-service-communicator feature will be responsible
for communication between any appc microservices which are
developed.

The appc dmaap micro service connector is part of this feature.
MessagingConnector.java is the main new class here.

Some code is copied from appc event listener and dmaap adapter.
Will be moved in a later commit.

Change-Id: Ic57f5ee6a9ab1538c6ddaa3e7c25ac0c9797fae0
Signed-off-by: Patrick Brady <patrick.brady@att.com>
Issue-ID: APPC-1744