Update patch set 2
Patch Set 2: Code-Review-2

Normally, we do want to use a logger in appc. But, in this CmdLine.java class, we actually do want to keep the System.out.println statements and not use a logger. Since the CmdLine.java class is meant to be run from a command line, we need the println statements to return messages to the user on the command line.

Patch-set: 2
Reviewer: Gerrit User 49 <49@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2